client/requirements.txt
certifi==2020.12.5
chardet==4.0.0
click==7.1.2
idna==2.10
numpy==1.19.5
Pillow==8.1.0
requests==2.25.1
urllib3==1.26.2
